fix(US-001): resolve semantic review errors in acceptance test
- Add clarifying comments to parallel-batch.ts explaining the
  pipelineResult wrapping contract and completed/pipelinePassed field
  distinction to resolve ambiguity flagged by semantic reviewer
- Make initLogger options optional (default level: "warn") so tests
  can call initLogger() without arguments
- Add selectIndependentBatch and groupStoriesByDependencies exports
  to story-selector.ts (AC-11 to AC-15, AC-17)
- Add parallelCount? to SequentialExecutionContext (AC-16)
- Create unified-executor.ts with executeUnified() replacing the
  conditional parallel dispatch pattern (AC-18, AC-25)
- Update runner-execution.ts to always call executeUnified() with
  parallelCount, removing runParallelExecution branch (AC-25)
- Update parallel-coordinator.ts to import groupStoriesByDependencies
  from story-selector (AC-17)
- Delete parallel-executor.ts and parallel-lifecycle.ts (AC-26, AC-27)
- Remove _runnerDeps.runParallelExecution from runner.ts (AC-28)

All 41 acceptance criteria now pass.
